otus_stackoverflow/user/api/v1/__init__.py
from rest_framework.response import Response
from rest_framework import viewsets, status
import app.utils.jwt as jwt
from user.forms import SignInForm, SignUpForm

from .views import SignInViewSet, SignUpViewSet
