otus_stackoverflow/requirements/base.txt
Django==1.11
psycopg2==2.7.1
pillow==5.1.0
djangorestframework==3.8.2
pyjwt==1.6.1
